// src/app/common/types/deal.type.ts
export interface Deal {
    id?: string | null;
    status?: 'open' | 'closed';
    openDate?: Date | null;
    closeDate?: Date | null;
    exchange?: string | null;
    feeType?: 'percent' | 'flat';
    feeOpen?: number | null;
    feeClose?: number | null;
    ticker?: string | null;
    assetType?: 'stock' | 'crypto' | 'option' | 'future';
    units?: number | null;
    openPrice?: number | null;
    invest?: number | null;
    openFeeAmount?: number | null;
    takeProfit?: number | null;
    takeProfitAmount: number | null;
    takeProfitFeeAmount?: number | null;
    takeProfitCostBasis?: number | null;
    takeProfitGross?: number | null;
    takeProfitPrice?: number | null;
    stopLoss?: number | null;
    stopLossAmount?: number | null;
    stopLossFeeAmount?: number | null;
    stopLossCostBasis?: number | null;
    stopLossGross?: number | null;
    stopLossPrice?: number | null;
    closeFeeAmount?: number | null;
    closeCostBasis?: number | null;
    closeGross?: number | null;
    closePrice?: number | null;
    profitLoss?: number | null;
    profitLossPercent?: number | null;
    progress?: number | null;
    progressPercent?: number | null;
    notes?: string | null;
    createdAt?: Date;
    updatedAt?: Date;
}